Integrated: Ported the failed unserialization test.
